delete reux file
